refacto: clean archi. and prepare impl. router
