+++ src/components/MainLayout.vue
@@ -1,0 +1,16 @@
+<template>
+  <tasks-layout />
+</template>
+
+<script>
+import TasksLayout from './Tasks/TasksLayout.vue'
+
+export default {
+  name: 'MainLayout',
+  components: {
+    TasksLayout
+  }
+}
+</script>
+
+<style scoped></style>
